+++ sass/02-base/_base.scss
@@ -56,6 +56,6 @@ html {
   width: 100%;
   height: 98%;
   background: var(--placeholder);
-  opacity: 1;
+  opacity: 0.7;
   transition: opacity 0.8s ease;
 }
