--- sass/04-components/_home.scss
@@ -1,154 +1,0 @@
-.homepage {
-  // padding: 122px 26px 88px;
-  background-color: $color_background;
-  // @media screen and (min-width: 576px) {
-  //   padding: 122px 32px 88px;
-  // }
-  &__list {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    row-gap: 8vh;
-    @media screen and (min-width: 992px) {
-      row-gap: 20vh;
-    }
-  }
-  &__items {
-    display: flex;
-    flex-direction: column;
-    grid-gap: 12px;
-    width: fit-content;
-    & span {
-      font-family: $rivalSans;
-      font-size: 14px;
-      color: var(--color);
-      line-height: 18px;
-      font-weight: $regular;
-    }
-    & h1 {
-      color: var(--color);
-      font-family: $sirenia;
-      font-size: 38px;
-    }
-  }
-  &__image {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: fit-content;
-    min-width: 100%;
-    & a {
-      position: relative;
-      display: block;
-      height: 100%;
-      width: fit-content;
-    }
-    & img {
-      position: relative;
-      width: 100%;
-      z-index: 2;
-      max-width: 600px;
-      height: auto;
-      object-fit: cover;
-      object-position: center;
-    }
-  }
-}
-
-.sliders {
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.slides {
-  height: calc(100% - 122px);
-  width: 100%;
-  position: relative;
-  // &__content {
-  //   height: 90%;
-  //   width: 100%;
-  //   position: relative;
-  //   min-height: 700px;
-  // }
-}
-
-.slide {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  position: absolute;
-  height: 100%;
-  width: 100%;
-  // grid-area: 1 / 1 / 2 / 2;
-  opacity: 0;
-  pointer-events: none;
-  width: 100%;
-  justify-content: center;
-  background-color: var(--color2);
-  &--active {
-    opacity: 1;
-    pointer-events: auto;
-  }
-  & h1 {
-    font-family: $tuskerGrotesk;
-    font-size: 64px;
-    text-transform: uppercase;
-    color: var(--color);
-    margin-bottom: 76px;
-    // letter-spacing: 0.03em;
-  }
-  & a {
-    font-family: $rivalSans;
-    font-weight: $bold;
-    font-size: 14px;
-    text-decoration-line: underline;
-    text-transform: uppercase;
-    color: var(--color);
-  }
-}
-
-.slideshow__controls {
-  height: auto;
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: baseline;
-  column-gap: 22px;
-  padding: 0 26px;
-  max-width: 470px;
-  & button {
-    background: none;
-    cursor: pointer;
-  }
-}
-
-.slideshow__controls__container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  column-gap: 8px;
-  width: 100%;
-}
-
-.slides__thumbs {
-  width: 100%;
-  max-width: 50px;
-  height: 30px;
-  border-radius: 5px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  opacity: 0.5;
-  & img {
-    max-height: 15px;
-  }
-  &--active {
-    opacity: 1;
-  }
-}
