+++ sass/04-components/_projects.scss
@@ -401,16 +401,12 @@
       transition: clip-path 0.5s ease-in-out;
       margin-bottom: 44px;
       max-height: 450px;
-      transform: scale(1.1);
-      transition: transform 0.8s ease, opacity 0.5s ease;
       opacity: 0;
       &--loaded {
         transform: scale(1);
         opacity: 1;
       }
-      @media screen and (min-width: 992px) {
-        max-width: 800px;
-      }
+      
     }
   }
   &__list {
@@ -465,6 +461,3 @@
   }
 }
 
-.slide__content {
-  overflow: hidden;
-}
