+++ sass/styles.scss
@@ -11,5 +11,5 @@
 @import './03-globals/menu';
 
 @import './04-components/projects';
-@import './04-components/home';
+@import './04-components/landing';
